# storage: delete mvccGetInternal, use mvccGet directly