This commit deletes a major chunk of legacy code in MVCC mutation code. `mvccPutInternal` was avoiding using mvccGet because this used to result in a cgo call. This led to a large amount of duplicate logic. Now that `mvccGet` calls into the `pebbleIterator` in Go, using it is much less expensive, so we can remove the re-implementation.

With a bit of tuning in follow-on commits, this results in a negligible performance hit which is well worth the elimination of duplicate code:
```
name                                                  old time/op    new time/op    delta
MVCCConditionalPut_Pebble/Create/valueSize=10-16        7.33µs ± 8%    7.04µs ± 7%     ~     (p=0.075 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=100-16       7.99µs ± 8%    7.89µs ± 8%     ~     (p=0.631 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=1000-16      8.51µs ± 7%    8.47µs ± 4%     ~     (p=0.529 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=10000-16     23.7µs ± 6%    23.8µs ± 5%     ~     (p=0.897 n=10+8)
MVCCConditionalPut_Pebble/Replace/valueSize=10-16       12.0µs ± 6%    12.0µs ± 7%     ~     (p=0.704 n=10+9)
MVCCConditionalPut_Pebble/Replace/valueSize=100-16      13.2µs ± 7%    13.0µs ± 9%     ~     (p=0.604 n=9+10)
MVCCConditionalPut_Pebble/Replace/valueSize=10000-16    28.8µs ± 7%    30.2µs ± 3%   +4.91%  (p=0.008 n=10+9)
MVCCConditionalPut_Pebble/Replace/valueSize=1000-16     11.5µs ± 6%    12.1µs ± 6%   +5.22%  (p=0.007 n=10+10)

name                                                  old speed      new speed      delta
MVCCConditionalPut_Pebble/Create/valueSize=10-16      1.37MB/s ± 8%  1.42MB/s ± 8%     ~     (p=0.085 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=100-16     12.5MB/s ± 7%  12.7MB/s ± 9%     ~     (p=0.645 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=1000-16     118MB/s ± 8%   118MB/s ± 4%     ~     (p=0.529 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=10000-16    423MB/s ± 6%   421MB/s ± 5%     ~     (p=0.897 n=10+8)
MVCCConditionalPut_Pebble/Replace/valueSize=10-16      835kB/s ± 5%   833kB/s ± 7%     ~     (p=0.734 n=10+9)
MVCCConditionalPut_Pebble/Replace/valueSize=100-16    7.56MB/s ± 7%  7.69MB/s ±10%     ~     (p=0.619 n=9+10)
MVCCConditionalPut_Pebble/Replace/valueSize=10000-16   347MB/s ± 7%   331MB/s ± 3%   -4.80%  (p=0.008 n=10+9)
MVCCConditionalPut_Pebble/Replace/valueSize=1000-16   87.1MB/s ± 6%  82.8MB/s ± 6%   -4.96%  (p=0.007 n=10+10)

name                                                  old alloc/op   new alloc/op   delta
MVCCConditionalPut_Pebble/Create/valueSize=10-16          366B ± 3%      366B ± 2%     ~     (p=1.000 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=100-16         909B ± 2%      905B ± 2%     ~     (p=0.492 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=1000-16      6.42kB ± 2%    6.46kB ± 1%     ~     (p=0.197 n=10+10)
MVCCConditionalPut_Pebble/Create/valueSize=10000-16     79.8kB ± 8%    77.7kB ±10%     ~     (p=0.190 n=10+10)
MVCCConditionalPut_Pebble/Replace/valueSize=1000-16     8.37kB ± 6%    9.05kB ± 4%   +8.20%  (p=0.000 n=10+9)
MVCCConditionalPut_Pebble/Replace/valueSize=10-16         390B ± 3%      435B ± 4%  +11.41%  (p=0.000 n=10+10)
MVCCConditionalPut_Pebble/Replace/valueSize=10000-16    72.0kB ± 7%    81.0kB ± 3%  +12.38%  (p=0.000 n=9+10)
MVCCConditionalPut_Pebble/Replace/valueSize=100-16      1.02kB ± 3%    1.16kB ± 2%  +14.05%  (p=0.000 n=10+10)

name                                                  old allocs/op  new allocs/op  delta
MVCCConditionalPut_Pebble/Create/valueSize=10-16          2.00 ± 0%      2.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Create/valueSize=100-16         2.00 ± 0%      2.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Create/valueSize=1000-16        2.00 ± 0%      2.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Create/valueSize=10000-16       3.00 ± 0%      3.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Replace/valueSize=10-16         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Replace/valueSize=100-16        4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Replace/valueSize=1000-16       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
MVCCConditionalPut_Pebble/Replace/valueSize=10000-16      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
```

If we want to continue pulling on this – 2 of the 4 remaining heap allocations here are due to the key copy in [`pebble.Iterator.SeekPrefixGE`](https://github.com/cockroachdb/pebble/blob/2da62788eedefce8dfbeef56a2324603606e18fb/iterator.go#L386). Caching this buffer across calls to `SeekPrefixGE` seems trivial. Caching the buffer across other method calls seems trickier due to the way we use the existence of the slice to indicate "prefix iteration". Maybe someone closer to this code will have good ideas about what to do about this.